+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### 0.0.3 (2017-04-15)
 
-1. add "Goto Nginx Docuement" context menu to watching directives and variables document
+1. add "Goto Nginx Document" context menu to watching directives and variables document
 2. filter directive completion items according to context(block)
 3. add more useful snippet
 4. fixed directives and variables completion item order